Fix linux compile issue.  Patch by Adam Treat <treat@kde.org>. Original review at http://codereview.chromium.org/3170
Review URL: http://codereview.chromium.org/4043

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@2494 0039d316-1c4b-4281-b951-d872f2087c98

+++ webkit/glue/autocomplete_input_listener.cc
@@ -64,9 +64,9 @@ bool HTMLInputDelegate::IsCaretAtEndOfText(size_t input_length,
   // friendliest behavior we can offer. Once the bug is fixed this method
   // should no longer need the previous_length parameter.
   if (((element_->selectionEnd() != element_->selectionStart() + 1) ||
-       (element_->selectionEnd() != input_length)) &&
+       (element_->selectionEnd() != static_cast<int>(input_length))) &&
       ((element_->selectionEnd() != element_->selectionStart()) ||
-       (element_->selectionEnd() != previous_length))) {
+       (element_->selectionEnd() != static_cast<int>(previous_length)))) {
     return false;
   }
   return true;
